Local-Node/local-node-backend/requirements.txt
uvicorn==0.34.2
fastapi==0.115.12
requests==2.32.3
python-jose[cryptography]==3.3.0 
dotenv==0.9.9
pandas==2.2.3
pymongo==4.13.0
python-multipart==0.0.20
openpyxl==3.1.5